fix: gpu inits are too slow => use SSD model
ssd_mobilenet_v1.pb object detector from tensorflow model zoo
is now used in place of faster_rcnn_inception_v2.pb because
it is a smaller model, faster to load and with faster
inference time. It is twice faster in both cases.

As a consequence to detect as many good bounding boxes as faster_rcnn
the cutoff threshold has been significantly reduced. This generates some
additional false bounding boxes detection but they are removed by the
2nd stage of RED/NOT RED classification.

Also since Udacity is now using a polarization filter in front of the
camera, when the traffic light is on, the pixels are now viewed as very
white. So the HLS classifier has been updated to deal with this
polarization filter.